requirements.txt
discord.py
playwright
aiohttp
python-dotenv